Add files via upload
This assignment introduces the basics of PLC programming using CODESYS. The goal is to learn programming techniques for controlling cylinders and to familiarize yourself with essential PLC tools. This foundational task emphasizes solving problems independently and mastering "learning by failing.
The code was tested on real cylinders at the Chalmers laboratory in Lindholmen.